+++ CHANGELOG.md
@@ -17,3 +17,4 @@ chains on the hexagonal Matrix.
 the scope handles for access to it's capture buffers.
 * Feature: Added WBlockDSP visual programming language utilizing the `synfx-dsp-jit` crate.
 * Feature: Added the `FormFM` node that was contributed by Dimas Leenman (aka Skythedragon).
+* Feature: Added `MidiP` node for MIDI pitch/note input.